general fixes
